Create an option to control whether $_SESSION would be used or not
There are two reasons where this option could need to be false.

1. When sessions are disabled. This may be cli programs where the
`$_SESSION` variable is undefined or just a request with no session
started.

2. During testing. One may want to unit test his application with
different languages. "Remembering" the language from the first run and
then using it for all further requests makes this impossible.

The original behaviour is preserved and saving to the session is
the default behaviour. One have to explicitly add the option
`saveToSession = false` to change it.